Move segment logic to ChartModel+Segment.swift
Refactored segment-related methods and viewport alignment from ChartModel.swift into a new ChartModel+Segment.swift file for better code organization and separation of concerns.
